+++ rotkehlchen/api/server.py
@@ -132,7 +132,6 @@ from rotkehlchen.api.v1.resources import (
     LocationAssetMappingsResource,
     LocationLabelsResource,
     LocationResource,
-    LoopringBalancesResource,
     ManuallyTrackedBalancesResource,
     MatchAssetMovementsResource,
     MessagesResource,
@@ -315,7 +314,6 @@ URLS_V1: URLS = [
     ('/blockchains/eth/modules/<string:module>/balances', EvmModuleBalancesResource),
     ('/blockchains/eth/modules/<string:module>/v<string:version>/balances', EvmModuleBalancesWithVersionResource),  # noqa: E501
     ('/blockchains/eth/modules/<string:module>/stats', ModuleStatsResource),
-    ('/blockchains/eth/modules/loopring/balances', LoopringBalancesResource),
     ('/blockchains/evm/accounts', EvmAccountsResource),
     ('/blockchains/type/<string:chain_type>/accounts', ChainTypeAccountResource),
     ('/blockchains/<string:blockchain>/accounts', BlockchainsAccountsResource),