+++ rotkehlchen/data_migrations/manager.py
@@ -16,6 +16,7 @@ from rotkehlchen.data_migrations.migrations.migration_23 import data_migration_2
 from rotkehlchen.data_migrations.migrations.migration_24 import data_migration_24
 from rotkehlchen.data_migrations.migrations.migration_25 import data_migration_25
 from rotkehlchen.data_migrations.migrations.migration_26 import data_migration_26
+from rotkehlchen.data_migrations.migrations.migration_27 import data_migration_27
 from rotkehlchen.data_migrations.migrations.migrations_13 import data_migration_13
 from rotkehlchen.data_migrations.migrations.migrations_14 import data_migration_14
 from rotkehlchen.data_migrations.migrations.migrations_18 import data_migration_18
@@ -55,6 +56,7 @@ MIGRATION_LIST = [  # remember to bump LAST_USERDB_DATA_MIGRATION if editing thi
     MigrationRecord(version=24, function=data_migration_24),
     MigrationRecord(version=25, function=data_migration_25),
     MigrationRecord(version=26, function=data_migration_26),
+    MigrationRecord(version=27, function=data_migration_27),
 ]